Adding dependency checks to the Argo feature
