+++ lib/project_types/extension/models/types/subscription_management.rb
@@ -8,11 +8,11 @@ module Extension
         IDENTIFIER = 'SUBSCRIPTION_MANAGEMENT'
 
         def create(directory_name, context)
-          Models::Types::Argo.admin.create(directory_name, IDENTIFIER, context)
+          Features::Argo.admin.create(directory_name, IDENTIFIER, context)
         end
 
         def config(context)
-          Models::Types::Argo.admin.config(context)
+          Features::Argo.admin.config(context)
         end
       end
     end
